backend/requirements.txt
# Minimal pinned requirements — let pip resolve compatible versions
# Python 3.12 compatible

fastapi
uvicorn[standard]
python-multipart
email-validator
ujson
defusedxml

sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary
asyncpg

redis

langchain-core
langchain
langchain-openai
langchain-anthropic==0.1.23
langchain-community

anthropic>=0.30.0
openai>=1.30.0
google-generativeai

ragas==0.1.9
datasets

confluent-kafka==2.4.0
prometheus-fastapi-instrumentator

pydantic>=2.7.4
pydantic-settings
python-dotenv
httpx
tenacity

pytest
pytest-asyncio